factor out adding the defaults, validate generated warnings
