+++ introspection-engine/connectors/sql-introspection-connector/src/lib.rs
@@ -2,6 +2,7 @@ pub mod calculate_datamodel; // only exported to be able to unit test it
 mod commenting_out_guardrails;
 mod error;
 mod misc_helpers;
+mod prisma_1_defaults;
 mod sanitize_datamodel_names;
 mod schema_describer_loading;
 mod version_checker;
