+++ introspection-engine/connectors/sql-introspection-connector/tests/db_specific_introspection/add_prisma1_defaults/mod.rs
@@ -20,6 +20,9 @@ async fn add_cuid_default_for_postgres(api: &TestApi) {
         "#;
     let result = dbg!(api.introspect().await);
     custom_assert(&result, dm);
+
+    let warnings = dbg!(api.introspection_warnings().await);
+    assert_eq!(&warnings, "[{\"code\":5,\"message\":\"These id fields had a `@default(cuid())` added because we believe the schema was created by Prisma 1.\",\"affected\":[{\"model\":\"Book\",\"field\":\"id\"}]}]");
 }
 
 #[test_each_connector(tags("postgres"))]
@@ -34,11 +37,13 @@ async fn add_uuid_default_for_postgres(api: &TestApi) {
 
     let dm = r#"
             model Book {
-                id  String @default(cuid()) @id
+                id  String @default(uuid()) @id
             }
         "#;
     let result = dbg!(api.introspect().await);
     custom_assert(&result, dm);
+    let warnings = dbg!(api.introspection_warnings().await);
+    assert_eq!(&warnings, "[{\"code\":6,\"message\":\"These id fields had a `@default(uuid())` added because we believe the schema was created by Prisma 1.\",\"affected\":[{\"model\":\"Book\",\"field\":\"id\"}]}]");
 }
 
 #[test_each_connector(tags("mysql"))]
@@ -58,6 +63,8 @@ async fn add_cuid_default_for_mysql(api: &TestApi) {
         "#;
     let result = dbg!(api.introspect().await);
     custom_assert(&result, dm);
+    let warnings = dbg!(api.introspection_warnings().await);
+    assert_eq!(&warnings, "[{\"code\":5,\"message\":\"These id fields had a `@default(cuid())` added because we believe the schema was created by Prisma 1.\",\"affected\":[{\"model\":\"Book\",\"field\":\"id\"}]}]");
 }
 
 #[test_each_connector(tags("mysql"))]
@@ -77,4 +84,7 @@ async fn add_uuid_default_for_mysql(api: &TestApi) {
         "#;
     let result = dbg!(api.introspect().await);
     custom_assert(&result, dm);
+
+    let warnings = dbg!(api.introspection_warnings().await);
+    assert_eq!(&warnings, "[{\"code\":6,\"message\":\"These id fields had a `@default(uuid())` added because we believe the schema was created by Prisma 1.\",\"affected\":[{\"model\":\"Book\",\"field\":\"id\"}]}]");
 }
